+++ src/Detail.js
@@ -19,7 +19,7 @@ import {
 import {
   getActions, getModelAttribute, getField,
   getDetailFields, getHasIndex, getModelLabel, getFieldLabel,
-  getFieldConditions, getHideable
+  getFieldConditions, getCollapsable
 } from './utils/schemaGetters'
 import { RecursiveTab } from './Tabs'
 import { getType } from './utils/getType'
@@ -55,12 +55,12 @@ const LabelInfoPopover = ({ LabelInfoComponent, fieldLabel }) => (
   />
 )
 
-export const HideTableButton = ({ modelName, fieldName, id, hideTable, hideTableChange }) => {
-  const HideTableIcon = hideTable ? FaAngleRight : FaAngleDown
+export const CollapseTableButton = ({ modelName, fieldName, id, collapse, collapseTableChange }) => {
+  const CollapseTableIcon = collapse ? FaAngleDown : FaAngleDown
   return (
-    <HideTableIcon
-      className={`hide-icon-${hideTable ? 'angle-right' : 'angle-down'}`}
-      onClick={() => hideTableChange({ modelName, fieldName, id, hideTable })}
+    <CollapseTableIcon
+      className={`hide-icon-${collapse ? 'angle-right' : 'angle-down'}`}
+      onClick={() => collapseTableChange({ modelName, fieldName, id, collapse })}
     />
   )
 }
@@ -236,17 +236,17 @@ export const DefaultDetailTableTitleWrapper = ({ children }) => {
   )
 }
 
-export const DefaultDetailO2MTableTitle = ({ schema, modelName, fieldName, id, targetInverseFieldName, targetModelName, path, node, user, hideable, hideTable, hideTableChange, customProps }) => {
+export const DefaultDetailO2MTableTitle = ({ schema, modelName, fieldName, id, targetInverseFieldName, targetModelName, path, node, user, collapsable, collapse, collapseTableChange, customProps }) => {
   const creatable = isCreatable({ schema, modelName: targetModelName, parentNode: node, user, customProps })
 
   return (
     <DefaultDetailTableTitleWrapper>
-      {hideable && <HideTableButton {...{
+      {collapsable && <CollapseTableButton {...{
         modelName,
         fieldName,
         id,
-        hideTable,
-        hideTableChange
+        collapse,
+        collapseTableChange
       }}/>}
       <DefaultDetailLabel {...{ schema, modelName, fieldName, node, customProps }} />
       { creatable && <DetailCreateButton {...{
@@ -270,9 +270,9 @@ const DefaultDetailM2MTableTitle = ({
   path,
   targetModelName,
   user,
-  hideable,
-  hideTable,
-  hideTableChange,
+  collapsable,
+  collapse,
+  collapseTableChange,
   customProps
 }) => {
   const editable = isFieldEditable({ schema, modelName, fieldName, node, user, customProps })
@@ -280,12 +280,12 @@ const DefaultDetailM2MTableTitle = ({
   return (
     <div style={{ marginBottom: '10px' }}>
       <h4 className='d-inline'>
-        {hideable && <HideTableButton {...{
+        {collapsable && <CollapseTableButton {...{
           modelName,
           fieldName,
           id,
-          hideTable,
-          hideTableChange
+          collapse,
+          collapseTableChange
         }}/>}
         {getFieldLabel({ schema, modelName, fieldName, node, customProps })}
       </h4>
@@ -359,9 +359,9 @@ export const DefaultDetailTable = ({
   const onDelete = R.path(['delete', 'onDetailDelete'], actions)
   const onEditSubmit = R.path(['edit', 'onDetailTableEditSubmit'], actions)
   const type = getType({ schema, modelName, fieldName })
-  const hideTable = R.path(['hideTable', modelName, id, fieldName], tableView)
-  const hideTableChange = R.path(['tableOptions', 'hideTableChange'], actions)
-  const hideable = getHideable(schema, modelName, fieldName)
+  const collapse = R.path([modelName, 'fields', fieldName, 'collapse'], tableView)
+  const collapseTableChange = R.path(['tableOptions', 'collapseTableChange'], actions)
+  const collapsable = getCollapsable(schema, modelName, fieldName)
 
   if (!data) { return <div className='container'>Loading...</div> }
 
@@ -383,9 +383,9 @@ export const DefaultDetailTable = ({
           path,
           targetModelName,
           user,
-          hideable,
-          hideTable,
-          hideTableChange,
+          collapsable,
+          collapse,
+          collapseTableChange,
           customProps
         }} />
         }
@@ -412,7 +412,7 @@ export const DefaultDetailTable = ({
             fieldOrder,
             user,
             tableView,
-            hideTable,
+            collapse,
             modalData,
             customProps
           }}
@@ -496,9 +496,9 @@ export const DefaultDetailTable = ({
           path,
           targetModelName,
           user,
-          hideable,
-          hideTable,
-          hideTableChange,
+          collapsable,
+          collapse,
+          collapseTableChange,
           customProps
         }} /> }
         { skipOverride(ValueOverride) ? null : <DetailValue
@@ -524,7 +524,7 @@ export const DefaultDetailTable = ({
             fieldOrder,
             user,
             tableView,
-            hideTable,
+            collapse,
             modalData
           }}
         /> }