+++ CHANGES.md
@@ -1,8 +1,12 @@
 # Conveyor Changelog
 
-## Version 1.5.0
+## Version 1.6.0
 
 ### Unreleased
+
+## Version 1.5.0
+
+### Released 1/30/20
 
 -   Add singleton feature #113
 -   Add disabled options dropdown #111
@@ -14,6 +18,7 @@
 -   Changed Tooltip action model to modelName #122
 -   Fix filter dropdown bug #124
 -   Give filter, hide, and sort icons unique classnames #120
+-   Add redirect to PageNotFound with bad url #123
 -   Add pagination #126
 -   Hide table headers if empty #115
 -   Filterable and Sortable are bool or func; currency filterable now #129
@@ -22,6 +27,8 @@
 -   Text area now correctly display new lines #132
 -   Use react-icons instead of react-svg #125
 -   Set clearable to false if required, other set to isClearable #100
+-   Refactor 'inputTypes', delete 'getInputType' #138
+-   Refactor tableView structure to fit pagination #135
 
 ## Version 1.4.0